moved constants to types for each component
