--- src/core/__mocks__/constants.ts
@@ -1,1 +1,0 @@
-export const radioGroupKey = 'radioGroup'
