+++ src/core/components/ga-dropdown/types.ts
@@ -1,10 +1,5 @@
 import type { InjectionKey, Ref } from 'vue'
 
-export interface RadioGroupContext {
-  name: string
-  model: Ref<string>
-}
-
 export interface DropdownChildPayload {
   id: string // ID is essential for aria-activedescendant
   value?: string | number
@@ -19,5 +14,4 @@ export interface DropdownContext {
   model?: Ref<string | number | undefined>
 }
 
-export const radioGroupKey: InjectionKey<RadioGroupContext> = Symbol('radioGroup')
 export const dropdownKey: InjectionKey<DropdownContext> = Symbol('dropdown')
