+++ src/core/components/ga-radio/types.ts
@@ -1,0 +1,8 @@
+import type { InjectionKey, Ref } from 'vue'
+
+export interface RadioGroupContext {
+  name: string
+  model: Ref<string>
+}
+
+export const radioGroupKey: InjectionKey<RadioGroupContext> = Symbol('radioGroup')
